rename 1d and 2d path effects to keep the sort order in the filesystem
