+++ skia-safe/src/effects/_1d_path_effect.rs
